requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
pandas
python-dotenv
Flask-Assets==2.0
pytailwindcss==0.1.4
resume-parser